Rename and relocate files
